requirements.txt
loguru
py-cord
pyyaml
python-dotenv
requests
